# docs/swagger.yaml
definitions:
...
  filler_internal_dto.BindParamRequest:
    properties:
      component_id:
        example: 10
        type: integer
      param_id:
        example: 5
        type: integer
    required:
    - component_id
    - param_id
    type: object
...
  filler_internal_dto.ConfigurationCreate:
    properties:
      device_component_id:
        example: 4
        type: integer
      indicator_id:
        example: 2
        type: integer
    required:
    - indicator_id
    type: object
  filler_internal_dto.ConfigurationUpdate:
    properties:
      device_component_id:
        example: 5
        type: integer
      indicator_id:
        example: 2
        type: integer
    required:
    - indicator_id
    type: object
...
  filler_internal_model.Configuration:
    properties:
      device_component:
        $ref: '#/definitions/filler_internal_model.DeviceComponent'
      id:
        example: 1
        type: integer
      indicator:
        $ref: '#/definitions/filler_internal_model.DeviceIndicator'
      thresholds:
        items:
          $ref: '#/definitions/filler_internal_model.Threshold'
        type: array
    type: object
...
  filler_internal_model.DeviceComponent:
    properties:
      components:
        items:
          $ref: '#/definitions/filler_internal_model.DeviceComponent'
        type: array
      id:
        example: 1
        type: integer
      internal_order:
        example: 1
        type: integer
      mappings:
        items:
          $ref: '#/definitions/filler_internal_model.Mapping'
        type: array
      model_id:
        example: 10
        type: integer
      parent_id:
        example: 3
        type: integer
    type: object
  filler_internal_model.DeviceIndicator:
    properties:
      contact:
        example: sysadmin@company.com
        type: string
      description:
        example: 'Hardware: x86 family'
        type: string
      id:
        example: 1
        type: integer
      location:
        example: Server Room, Rack 3
        type: string
      name:
        example: Core-Switch-01
        type: string
      object_id:
        example: 1.3.6.1.4.1.9.1.516
        type: string
      services:
        example: 12
        type: integer
    type: object
  filler_internal_model.Mapping:
    properties:
      coefficient:
        example: "1.500000000000"
        type: string
      enum:
        type: object
      frequency:
        example: MEDIUM
        type: string
      id:
        example: 1
        type: integer
      indicator_id:
        example: 2
        type: integer
      param_id:
        example: 5
        type: integer
    type: object
...
  filler_internal_model.Threshold:
    properties:
      enabled:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      level:
        example: WARNING
        type: string
      operator:
        example: ==
        type: string
      prev_operator:
        example: '&&'
        type: string
      previous_id:
        example: 2
        type: integer
      previous_threshold:
        $ref: '#/definitions/filler_internal_model.Threshold'
      source_internal_order:
        example: 1
        type: integer
      source_model:
        example: 10
        type: integer
      source_param:
        example: temperature
        type: string
      target_device:
        example: 5
        type: integer
      target_param:
        example: alarm_state
        type: string
      type:
        example: INT
        type: string
      value:
        example: "85"
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: victor-14-244@mail.ru
    name: Lorgar Avrelian
    url: https://github.com/Lorgar-Avrelian
  description: HTTP-сервер на Go с автоматической генерацией Swagger при старте
  license:
    name: Apache 2.0
    url: http://apache.org
  title: DB Filler API
  version: "1.0"
paths:
...
  /api/v1/configurations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/filler_internal_model.Configuration'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все рабочие конфигурации
      tags:
      - '10. Конфигурация: Конфигурации устройств'
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные конфигурации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/filler_internal_dto.ConfigurationCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Возвращает полностью раскрытое содержание созданной конфигурации
          schema:
            $ref: '#/definitions/filler_internal_model.Configuration'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать рабочую конфигурацию
      tags:
      - '10. Конфигурация: Конфигурации устройств'
  /api/v1/configurations/{id}:
    delete:
      parameters:
      - description: ID Конфигурации
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить рабочую конфигурацию по ID
      tags:
      - '10. Конфигурация: Конфигурации устройств'
    get:
      parameters:
      - description: ID Конфигурации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает полностью раскрытое содержание
          schema:
            $ref: '#/definitions/filler_internal_model.Configuration'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить рабочую конфигурацию по ID
      tags:
      - '10. Конфигурация: Конфигурации устройств'
    put:
      consumes:
      - application/json
      parameters:
      - description: ID Конфигурации
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/filler_internal_dto.ConfigurationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает полностью раскрытое обновленное содержание
          schema:
            $ref: '#/definitions/filler_internal_model.Configuration'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить рабочую конфигурацию по ID
      tags:
      - '10. Конфигурация: Конфигурации устройств'
  /api/v1/configurations/bind:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID конфигурации и ID порога (используем BindParamRequest для
          совместимости структуры)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/filler_internal_dto.BindParamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Привязать порог к рабочей конфигурации
      tags:
      - '10. Конфигурация: Конфигурации устройств'
  /api/v1/configurations/bind/{configurationId}/{thresholdId}:
    delete:
      parameters:
      - description: ID Рабочей конфигурации
        in: path
        name: configurationId
        required: true
        type: integer
      - description: ID Порога
        in: path
        name: thresholdId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить связь рабочей конфигурации с порогом
      tags:
      - '10. Конфигурация: Конфигурации устройств'
...
swagger: "2.0"
...